requirements.txt
requests
PyYAML
rich
textual